Make ArrayList throw exceptions as specified in JDK Javadoc (#878)
* Throw more exceptions as specified in the JDK Javadoc

Array[E].toArray[T](Array[T]) is not still compliant because
it is specified to throw ArrayStoreException if not T >: E but
it doesn't.

This can't be fixed now because there is no compile-time type
information and runtime cast is not type-checked. See #858.

* Change the type of ArrayList.inner to Array[Any]

Found that Scala(-Native) can handle Array[Any] just fine.