Build API  (#1143)
* Extract `discover` to new `llvm` package in tools

* Extract `nativeCompileOptions` to `llvm` package

They're now defined under the name `defaultCompileOptions`.

* Extract `defaultLinkingOptions` to `llvm` package

* Extract target detection out of sbt plugin

* Extract `unpackNativeLib` out of the sbt plugin

This commit adds a new package `scala.scalanative.build`. The expected
users of this package are build tools that want to implement support for
Scala Native.

* Extract `nativeLinkNIR` out of sbt plugin

* Extract `nativeCompileLL` out of sbt plugin

* Move `GarbageCollector` to `tools` module

* Extract `nativeCompileLib` out of sbt plugin

* Extract `nativeLinkLL` out of sbt plugin

* Move to java.nio

* Add more info to `Config`

The following fields are added:
 - `clang` and `clangpp`
 - `linkingOptions` and `compileOptions`
 - `gc`

 These informations are often passed around in our toolchain. It helps
 to define cleaner API if we just pass around a `config` object instead
 of many parameters.

* Put `linkNIR` back in sbt plugin

All the important work is already done in `tools.link`. The rest was
just logging of information, which should be left to the user of our
API.

* Fix `IO.deleteRecursive`

Trying to delete a non-existing directory would throw an exception.

* Add `IO.write` to write to a file

Very similar to `Files.write`, but create the parent directories before.

* Fix calls to `IO.getAll`

* Add `build` to `build`

This function does all the steps from NIR to executable binary.

* Fix scripted test `optimizer-reporter`

Broken by the switch to NIO.

* Add `nativeLib` to `Config`

This simplifies the API.

* Configure `Driver` with `Mode` instead of `Config`

We would like to add `Driver` as part of `Config`. This is to avoid
circular dependency.

* Add `name` to `Mode` and `default`

The `default` method in `object Mode` describes the default value for
the `Mode`. It is currently set to `Debug`, following what was
previously done.

`Mode` instances also get a new attribute `name` which is a string
representation of the `Mode`.

Finally, add `apply` method to `Mode` to get an instance from a String.

* Add driver, {linker,Optimizer}Reporter to Config

* Add `Config.default`

This function takes a minimal number of arguments to create a config
with sane default parameters.

* Remove `mode` from `Config`, put into `Driver`

Otherwise, we may have `Config` instances where the `Driver` and the
`Mode` disagree.

* Store `entry` as `String` in `Config`

This hides the NIR from the public API.

* Fix `linkerReporter`

* Fix `LinkerSpec`

* Setup MiMA

This is currently set up to use MiMA 0.1.14, which is slightly outdated.
I had issues with more recent versions of MiMA (NoSuchMethodExceptions
etc., apparently because of compatibility layer with sbt 1 probably).

* Make `passes` and `withPasses` in `Driver` private

* Run MiMA in release script and CI

* Make methods of `Reporter` private

This holds for the linker and the optimizer.

* fixup mima

* Move `Config` to `build` package

* Merge tools and build namespaces

* Simplification

* Move `package object llvm` to `object LLVM`

* Fold `tools` inside `build`

* Run MiMa as part of `test-all`

* Run tests before releasing

* Move `logger` inside `Config`

* Use sbt's logger instead of the default one

* Move logging outside of sbt plugin

* Make `codegen` return generated paths

* Hide substeps in API

* Remove subtasks from sbt plugin

* Consume linking error in `build`

* s/ativeLib/ativelib/g

* Logger: add implementation for `running` and `time`

* s/GarbageCollector/GC/s

* Fix formatting

* Add documentation for `Logger`

* Add documentation for `Config`

* Add more doc

* Revive log about GC in use

* Improve documentation for `build.build`

* Throw `IllegalArgumentException` for unknown GCs

* Add `BuildException`