Improve portability of time.h bindings (#1170)
* Improve portability of time

The time.h specification only requires that the first nine integer
fields in the struct are present, but doesn't guarantee the order.
Moreover, many implementations add additional fields that they will
modify when calling the standard functions which can cause memory
corruption if only enough memory is allocated for the mandatory fields.
To address this, I introduce a new scalanative_tm struct that has the
nine mandatory fields and marshal that struct to and from the platform
specific native tm struct. This is is similar to what is done in stat.c.

A number of the methods are not thread safe because they return pointers
to native structs. To work around this, I only implement the _r variants
of these methods. To keep the api reasonably nice for the user, I added
versions of the non '_r' prefixed variants that require an implicit
Zone. These methods allocated the required data structures and pass them
into the '_r' variants. The returned pointers will have the same
lifecycle as the zone.

* Move native/time.scala to posix/time.scala

The time.scala file that added the methods in time.h was inconsistently
located in the native package. This commit moves it into the posix
package and also unifies the declaration of the time types. The time_t
and timespec types are declared in time.h, while timeval is declared in
sys/time.h. I removed the redundant declarations.